requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
aiohttp>=3.9.0
streamlit>=1.28.0
Pillow>=9.0.0
aiosqlite>=0.19.0
